Several small 'typo' type fixes.
Incremental checkin towards 'crack opening' demonstration.

Main work relates to development of 'crack' opening 
calculation.  The demonstration code is 
basically set up.  However, I forgot/overlooked 
the fact that the jacobian also has to reflect 
the presence of the crack.  Need to do a small 
amount of work on that aspect for this demonstration 
code to run properly.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@629 d9297710-12f9-482b-b11f-fb677dc88bbb

+++ pditi/aliti/utQuadrulator.cxx
@@ -24,7 +24,7 @@ static int myRank;
 void getGrid(){
 	double x0 = 0.0, y0 = 0.0;
 	size_t nx = 3, ny = 3;
-	size_t hx = 1.0, hy = 1.0;
+	double hx = 1.0, hy = 1.0;
 	Quadrulator reader (x0, y0, nx, hx, ny, hy);
 
 	ArrayRCP<size_t> links = reader.getVertexLinks();
